forgein keys fixing

+++ app/database/migrations/2014_06_21_020000_prod_difference.php
@@ -10,11 +10,10 @@ class ProdDifference extends Migration
     {
         Schema::create('prod_difference', function (Blueprint $table) {
 
-            $table->integer('id')->unsigned()->primary();
+            $table->smallInteger('id')->unsigned()->primary();
             $table->boolean('visible')->default(1);
             $table->tinyInteger('count')->unsigned();
             $table->string('name', 48)->unique();
-
             $table->engine = 'InnoDB';
         });
     }
